getting rid of int64_t

+++ 04-ext2-read-file/solution.c
@@ -15,8 +15,8 @@ int read_sb(int img, struct ext2_super_block* p_sb) {
 }
 
 int read_gd(int img, struct ext2_super_block* p_sb,
-            struct ext2_group_desc* p_gd, int64_t blk_sz, int inode_nr) {
-  int64_t gd_idx = (inode_nr - 1) / p_sb->s_inodes_per_group;
+            struct ext2_group_desc* p_gd, int blk_sz, int inode_nr) {
+  int gd_idx = (inode_nr - 1) / p_sb->s_inodes_per_group;
   off_t gd_offset = blk_sz * (p_sb->s_first_data_block + 1) +
                     gd_idx * sizeof(struct ext2_group_desc);
   if (pread(img, p_gd, sizeof(struct ext2_group_desc), gd_offset) == -1) {
@@ -27,8 +27,8 @@ int read_gd(int img, struct ext2_super_block* p_sb,
 
 int read_inode(int img, struct ext2_super_block* p_sb,
                struct ext2_group_desc* p_gd, struct ext2_inode* p_inode,
-               int64_t blk_sz, int inode_nr) {
-  int64_t inode_idx = (inode_nr - 1) % p_sb->s_inodes_per_group;
+               int blk_sz, int inode_nr) {
+  int inode_idx = (inode_nr - 1) % p_sb->s_inodes_per_group;
   off_t inode_offset =
       blk_sz * p_gd->bg_inode_table + p_sb->s_inode_size * inode_idx;
   if (pread(img, p_inode, sizeof(struct ext2_inode), inode_offset) == -1) {
@@ -37,17 +37,17 @@ int read_inode(int img, struct ext2_super_block* p_sb,
   return 0;
 }
 
-int read_blk(int img, int out, int32_t blk_idx, int32_t blk_sz,
-             int32_t* p_read_left, int level) {
+int read_blk(int img, int out, int blk_idx, int blk_sz, int* p_read_left,
+             int level) {
   int ret;
   if (level) {
-    int64_t ub = blk_sz / sizeof(int32_t);
-    int32_t* blk_idxs = malloc(blk_sz);
+    int ub = blk_sz / sizeof(int);
+    int* blk_idxs = malloc(blk_sz);
     if (pread(img, blk_idxs, blk_sz, blk_sz * blk_idx) == -1) {
       free(blk_idxs);
       return -errno;
     }
-    for (int64_t i = 0; i < ub; i++) {
+    for (int i = 0; i < ub; i++) {
       if (!blk_idxs[i]) {
         break;
       }
@@ -78,7 +78,7 @@ int dump_file(int img, int inode_nr, int out) {
   if ((ret = read_sb(img, &sb))) {
     return ret;
   }
-  int32_t blk_sz = DEFBLKSZ << sb.s_log_block_size;
+  int blk_sz = DEFBLKSZ << sb.s_log_block_size;
 
   if ((ret = read_gd(img, &sb, &gd, blk_sz, inode_nr))) {
     return ret;
@@ -88,7 +88,7 @@ int dump_file(int img, int inode_nr, int out) {
     return ret;
   }
 
-  int32_t read_left = inode.i_size;
+  int read_left = inode.i_size;
   for (int i = 0; i < EXT2_NDIR_BLOCKS; i++) {
     if ((ret = read_blk(img, out, inode.i_block[i], blk_sz, &read_left, 0))) {
       return ret;
